pyproject.toml
[project]
name = "geometrycrafter"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "numpy<2",
    "torch==2.9.1",
    "torchvision==0.24.1",
    "torchaudio==2.9.1",
    "xformers==0.0.33.post2",
    "triton-windows>=3.6.0.post25",
    "diffusers>=0.36.0",
    "transformers>=5.1.0",
    "accelerate>=1.12.0",
    "fire>=0.7.1",
    "imageio-ffmpeg>=0.6.0",
    "scipy",
    "opencv-python",
    "timm",
    "einops",
    "omegaconf",
    "decord>=0.6.0",
    "tqdm>=4.67.3",
    "huggingface-hub>=1.4.1",
    "kornia>=0.8.2",
    "opencv-contrib-python>=4.11.0.86",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
torchaudio = { index = "pytorch" }
xformers = { index = "pytorch" }
